requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
PyJWT==2.10.1
redis==5.2.1
SQLAlchemy==2.0.38
typing_extensions==4.12.2
Werkzeug==3.1.3
gunicorn==23.0.0